Integrated-Data-Platform-backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
joblib==1.2.0
numpy==2.3.3
openai==2.3.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tzdata==2025.2